Preparing for 2DRenderer - Moved EntryPoint.h file from Engine.h to SandboxApp. (It causes two mains funcs being defined at the same time sometimes.) -Changed the class "ImGuiLayer" to "ImGuiBase" that I now use only for initialization purposes. -LayerStack class is now a propoer stack instead of a queue. -Created a third layer in SandboxApplicatino.cpp. -Cleaned the SandboxApp class a little bit in "SandboxApplicatino.cpp". Nothing that is primitive related is done inside this class anymore. In other words, moved everything to their respective layers. -IMPORTANT: Every layer now can do it's own ImGui rendering however it wants inside "OnImGuiRender" functon.

+++ AnorEngine/Source/Events/Event.h
@@ -108,8 +108,8 @@ namespace AnorEngine
 			}
 		public:
 			void Log() override { INFO("Mouse Move Event. Mouse Position: X: {0}, Y: {1}", m_MouseXPosition, m_MouseYPosition); }
-			int GetMouseXPosition() { return m_MouseXPosition; }
-			int GetMouseYPosition() { return m_MouseYPosition; }
+			float GetMouseXPosition() { return m_MouseXPosition; }
+			float GetMouseYPosition() { return m_MouseYPosition; }
 		public:
 			virtual ~MouseMoveEvent() override {}
 		private:
@@ -127,8 +127,8 @@ namespace AnorEngine
 			}
 		public:
 			void Log() override { INFO("Mouse Scroll Event."); }
-			int GetXOffset() { return m_MouseXOffset; }
-			int GetYOffset() { return m_MouseYOffset; }
+			float GetXOffset() { return m_MouseXOffset; }
+			float GetYOffset() { return m_MouseYOffset; }
 		private:
 			float m_MouseXOffset, m_MouseYOffset;
 		};
