+++ AnorEngine/Source/Events/EventHandler.cpp
@@ -13,13 +13,13 @@ namespace AnorEngine
 		{
 			return glfwGetKey(EventHandler::s_TargetApplication->GetOpenGLWindow()->GetNativeWindow(), keyCode); //GLFW_RELEASE equals to 0 thats why this works.
 		}
-		float EventHandler::GetMouseXOffset(GLFWwindow* windowPtr)
+		double EventHandler::GetMouseXOffset(GLFWwindow* windowPtr)
 		{
 			double xpos, ypos;
 			glfwGetCursorPos(windowPtr, &xpos, &ypos);
 			return xpos;
 		}
-		float EventHandler::GetMouseYOffset(GLFWwindow* windowPtr)
+		double EventHandler::GetMouseYOffset(GLFWwindow* windowPtr)
 		{
 			double xpos, ypos;
 			glfwGetCursorPos(windowPtr, &xpos, &ypos);
